проектная работа 11

+++ src/utils/api.js
@@ -44,24 +44,22 @@ class Api {
         .then((res) => this._checkResponse(res))
     }
 
-    //Поставить лайк карточке
+    //Поставить/удалить лайк карточке
 
-    addLikeCard(cardId) {
-        return fetch(`${this._url}/cards/likes/${cardId}`, {
-            method: 'PUT',
-            headers: this._headers,
-        })
-        .then((res) => this._checkResponse(res))
-    }
-
-    //Удалить лайк карточки
-
-    deleteLikeCard(cardId) {
-        return fetch(`${this._url}/cards/likes/${cardId}`, {
-            method: 'DELETE',
-            headers: this._headers,
-        })
-        .then((res) => this._checkResponse(res))
+    changeLikeCardStatus(cardId, isLiked) {
+        if (isLiked) {
+            return fetch(`${this._url}/cards/likes/${cardId}`, {
+                method: 'PUT',
+                headers: this._headers,
+            })
+            .then((res) => this._checkResponse(res))
+        } else {
+            return fetch(`${this._url}/cards/likes/${cardId}`, {
+                method: 'DELETE',
+                headers: this._headers,
+            })
+            .then((res) => this._checkResponse(res))
+        }
     }
 
     //Получить данные пользователя
@@ -87,7 +85,6 @@ class Api {
 
     //Отредактировать аватар пользователя
 
-    //editAvatarUser(avatar) {
     editAvatarUser({ avatar }) {    
         return fetch(`${this._url}/users/me/avatar`, {
             method: 'PATCH',
